First demo application for PyBossa. Closes #10 Closes #23
The first demo application shows to the user an image and asks the user
if he can see a human in the photo. Possible answers are: Yes, No, I
don't know.

flickerPerson/getPhotos.py is a script for creating tasks for PyBossa.
static/flickrPerson/js/flickrperson.js is a JS script to get the tasks
using the API from PyBossa.

index.html has been redirected to the example app, so it is easy to test
it for new users.