Dropdown for the applications. Closes #27
The application now adds a dropdown item to the topbar based on how many
applications are registered in the system, and if they are visible or
not. The stats['app'] variable has been also updated to reflect this
case, as that information will be shown to everyone (including anonymous
users), so hidden applications should not be counted.

The version of the application has been added to.