FlickrPerson updated for the new DB. Closes #37 Refs #24
The new DB model #24 included several changes to PyBossa. Basically, the
removal of the Batches table, and some new fields in some tables
(check the model.py for further details).

The FlickrPerson application finally saves the state of the tasks, so
this new version only presents un-completed task to users. This is
done via the API #5.