API has been secured. Refs #42
The API has been secured, using a decorator @authenticate defined in the
file util.py.

The decorator checks if there is an argument in the URL **api_key**, and
if the submitted api_key belongs to one of our users. If the answer is
yes, the API method (CRUD) will be run.

This is a similar approach to Flickr API security facilities.