Fix for the key error in stats.py Closes #73
When an application has been deleted and there is no correct clean up (truncate or delete in cascade) there might be some ghost task_runs in the DB causing the failure. This commit checks if the app.id exists in the DB or not before proceeding to get the info for that app.