Fix absolute paths using url_for() method. Closes #85
The templates have absolute paths breaking the application when deployed into
a folder and not a root server. This commit fixes the bug and improves the code.

Additionally, the index page for all the blueprints have been renamed to index()
to reflect that.

app_details() has been renamed to details().