Twitter sign in added. Closes #14
This commit adds support for Twitter OAuth sign in methods. This feature can be
enabled or disabled via the local settings (check the template file).

This commit introduces a change into the DB, so Alembic migration tool has been
used to support the migration of the schema. The migration does not support SQLite
as this DB does not fully support the ALTER TABLE sql instruction (check the documentation
of Alembic).

If you are running an old version of the DB use the alembic command to upgrade to head:

alembic upgrade head

Otherwise create the DB as usual, as the new field for the Twitter account has been
included in the model.