The Register tab has been removed. Closes #102
This commit improves the Sign in/out procedure, reducing the entry points and
providing a new an better layout for the users.

There is right now one entry point: Sign in, and from there you can Sign in with
your Twitter or PyBossa account. Additionally we give the option to create an
account too.

Finally, in order to provide a better layout this commit has updated Bootstrap to
version 2.0.4. The performance seems to have improved too.