Unit Tests for the Scheduler. Refs #109
This commit includes the Unit Tests for the new scheduler. As some tests for the
scheduler take really long, we have reduce the required amount of TaskRuns to
only 10. This will be fixed in the next commit, as via the Task creation it will
be possible to change the 30 limit.