Facebook sign in added. Closes #119
This commit adds support for Facebook sign in, following the same pattern used
in Twitter. The only main difference is that in order to get the user info
the oauth_token has to be stored in the session as we need to run a get('/me')
request. This request need the oauth_token, otherwise it will fail.