New /app/slug/task/task_id endpoint. Closes #127
Now users can access specific tasks by typing in the URL the task
ID. This will allow to use external forum services like Disqus (Refs #133)
and in the future to allow users to update their answer. For the moment,
if a user has already contributed to the task, a warning message will be shown, not
rendering the app presenter.

Using this new type of endpoint application owners will have to include within
the task.info JSON object a new field named question to render it in the presenter.

Please, check the app-flickrperson code for further details.