Fix #138 another Twitter 500 Error.
This commit fixes the problem: when a Twitter user tries to sign in
the e-mail was configured to this keyword: "None". If two users do not change it,
the system fails giving a DB key Unique error.

This commit fixes the issue, assigning to every new Twitter user as e-mail the
username:

 * user.email_addr = user.name

Before loading any view, the system checks for this condition, and if it is true
warns the user with an error message.