Unit tests for the tutorial feature.
This commit adds tutorial support for applications. When PyBossa
detects that a user is accessing for the firs time an application
checks if the app has a tutorial. In that case, it will show the
tutorial, a web page, and from that moment the use will not see again
that page unless he clicks in the help button of the application.