New api endpoint to show user progress. Closes #149
This new commit provides a new API endpoint that allows app owners to show the
progress of the users in their applications.

The new API endpoint is really simple:

 * http://domain/api/app/<int:app_id>/userprogress or
 * http://domain/api/app/<short_name>/userprogress

Will return a JSON object with the total number of available tasks and the
number of tasks completed by the user (it checks if the user is anonymous or not).

This new endpoint should be integrated into PyBossa.JS so users do not have to
query the URL explicitly. Something like:

 * pybossa.userprogress( app ).done(function(data){});

Could be the format to get the progress and load it into the application presenter.