Adding n_answers to the model and improving the sched. Closes #161 Refs #187
The following commit has added the n_answers to the Task model. The code
supports the task.info.n_answers but it is marked in the code as DEPRECATED.

This commit also includes a better scheduler regarding how the state and
the tasks are delivered. I run the following tests:

 * An application, Flickr Person Finder, with 20,000 tasks
 * Request a task

With the previous scheduler, PyBossa returned a task after 3977.17365098 seconds,
while with the new one it returns it in 0.01 seconds. The reason is because the
previous scheduler loaded all the tasks and iterated over them. The new one loads
tasks in batches reducing the overhead.

Finally the documentation has been updated with this changes.