Better scheduler using SQLAlchemy. Refs #187 Refs #161
This commit basically uses a new approach to get tasks for users using only
SQLAlchemy magic and not loops as before (the previous sched only provided 10
tasks due to an internal bug).

This new scheduler also forces to remove SQLite from testing. The production
server always runs on PostgreSQL so this commit has moved all the tests to work
with a PostgreSQL DB. If you try to run the tests with SQLite it will fail.

The performance has been improved again over the latest scheduler, obtaining in
some cases times of 0.001 in comparison with previous commit: 0.01.