Move tasks page to a different endpoint. Closes #196
This commit simplifies the application welcome home page by moving
the tasks page to another endpoint: http://domain/app/short_name/tasks

The tasks can be accessed from the Summary of the application.

This commit also helps with the performance as the tasks are only loaded
when the user requests it, not by default.