Better integration of templates when creating an app. Closes #276
This commit adds the three available templates and documentation when a
user clicks in the **create** button. The new page include links to all
the templates' source code as well as the documentation.

Via a button you can also display the web form as well as a clear hint that by selecting this method you can use the CSV importer.