Top users and top and featured apps now cached. Closes #300
This commit moves the front page actions of getting the top users and apps,
as well as the featured ones to the util.py module in order to provide caching
mechanisms at function level. Thanks to this caching improvements, loading the
front page only causes 1 SQL query, while without it gets 21. The improvement is
really big.