Cache items instead of full views. Refs #303
This commit shows how we can do some caching at the application or user level
using the flask-cache module. Basically, each time a view requests a list of apps
the apps will be formatted (avoiding the lazy loading) so they can be saved in the
cache. Then, when the view is requested, the items will be requested checking first
if the app is in the cache or not. This method allows us to basically store several
applications in the cache and only update those ones that have been updated in the DB
(like adding more tasks, etc.)