Code re-factoring. Closes #303
This commit uses a cache per object, so the server will be putting a domain object in the cache (app or user) for the front page and applications index instead of the full view.

This helps to reduce the cache misses by actually having as many items in the cache as possible, not full views. The benefits are that if one item has to be removed, we will be only updating one element and not all the full view that could have 5 or more items, punishing the DB again with the queries.