New layout and logic for creating an app. Refs #327
The new layout basically, checks first if after the creation of the application (the three options have been removed) if the app has a task presenter or not.

If the application does not have a template, it will offer 4 template choices that will be automatically rendered for the user, so they can easily adapt them for their needs.