Fixes for new unit tests. Closes #667
This commit fixes the no more tasks bug. The solution has been to add a DB
trigger that automatically updates the task.state based on the number of
answers saved, when a post is commited.

This solution simplifies the code, plus reduces its complexity as now the
scheduler does not have to clean the candidate tasks, as they are already good
ones, because now the task.state is updated in real time, not when a user
requests a new task.