Full refactor for using Flask Factories in PyBossa
This is the first big commit doing a big refactor for using Flask Factories.

Basically, every module like Sentinel, Signer, etc. are loaded as classes in
the module pybossa.extensions following the Flask pattern of extensions.

This allows us to load the objects in pybossa.core and configure it lately
whenever we want, allowing us to have more flexibility.

Additionally this commit simplifies a lot pybossa.web as we have now a very
clean run.py file that runs a PyBossa server.

Finally, all the tests are going to be refactor to address the new factory
setup.

This issue will have more commits as there is work to be done on upgrading the
tests.