Support filter instead of filter_by for JSON data
This commit adds a work around for querying JSON data. SQLAlchemy needs to use
filter instead of filter_by to do queries, unless you do a CAST.

PostgreSQL accepts '"astring"' as JSON valid, so if you want to search you need
to do a cast before querying it. This commit fixes this by including an
exception and handling it properly.