Fix misaligned data in task/task_run CSV export
Previously, headers were written using the first
instance of an object. This caused misalignment
when tasks where imported that did not have the
exact same column headers.

The fix goes through all objects to construct
the headers to guarantee that all headers are
written to the CSV before any data is written.