Add a counter table.
The breadth scheduler needs to count all the task runs for each task in
order to send the task with the less number of answers. This implies
a LEFT OUTER JOIN as well as a count(*) making the query really
expensive for big projects. The solution would be to compute this
value every time a new task is created as well as a new task_run is
created.

This will be handled in a separated table, in order to avoid deadlocks.