Allow users to remove tasks that have participated.
This PR will enable users to use the fulltext search API and get
only the tasks in which the user (anon, auth or external UID) could
participate.

By default, it will return all the tasks for the given call, but if
you pass in &participated=1, this flag will remove from the results
all the tasks that the user has contributed to.