Remove redundant queries.
The removed code was too heavy. In Crowdcrafting for example it has to
compute around 2500 SQL queries to analyze all the projects, which is
too much. As we've now specific endpoints with pagination via JSON the
pybossa theme could query them to show the data that anyone wants.

This should make PYBOSSA much faster, specially for the front page.