Add clean project method.
This will allow to reset a project to its initial state.

You can run the command like this:

``` python

python cli.py clean_project project_id skip_tasks

```

Project_id should be the project ID (integer) and skip_tasks (Boolean)
True or False.

If you run the command without the skip_tasks argument, everything will
be deleted. If you want to keep the tasks, and remove all the taskruns
and results, as well as reset the stats of the project, pass the True
argument.