feat(webhooks): add cli command to create webhooks.
This PR allows to re-create all the missing webhooks of a project.

For example, when you have a project running for a few months, and later
on you want to start analyzing everything. All you have to do is to use
the cli.py client library to create them:

```python
python cli.py create_webhooks projectID
```

Where projectID is the projectID that you want to populate with. NOTE:
the script does not recreate previous webhooks, those ar respected.