feat: add command to migrate social accounts.
This command moves all the social accounts to local accounts. It
generates a password for the account, and removes its ID from the social
columnt. NOTE: it's important that all the users should have an email.
Twitter does not return the email from the account, so those users
without an email will not be able to recover their password.