chore!: drop support for Flask-Oauth (twitter, facebook and google) (#2010)
* chore!: drop support for Flask-Oauth (twitter, facebook and google)

Flask-Oauth is no longer mantained, and the best for PYBOSSA is to
remove it. Also, Twitter, Facebook and Google signin have not been used
as widely as the local account. Thus, removing that unused code is for
the best.

* chore: use default theme.

* fix: pin cryptography package.

* feat: add command to migrate social accounts.

This command moves all the social accounts to local accounts. It
generates a password for the account, and removes its ID from the social
columnt. NOTE: it's important that all the users should have an email.
Twitter does not return the email from the account, so those users
without an email will not be able to recover their password.

* chore: remove flask-oauth

* chore: use master branch.